Map with Navigation styles moved to separate file
